+++ Core/OfficeDevPnP.Core/Pages/CanvasSectionTemplate.cs
@@ -30,7 +30,28 @@
         /// Two columns, left one is 1/3, right one 2/3
         /// </summary>
         TwoColumnRight = 5,
-
+#if !SP2019
+        /// <summary>
+        /// One column + one vertical section column
+        /// </summary>
+        OneColumnVerticalSection = 6,
+        /// <summary>
+        /// Two columns of the same size + one vertical section column
+        /// </summary>
+        TwoColumnVerticalSection = 7,
+        /// <summary>
+        /// Three columns of the size + one vertical section column
+        /// </summary>
+        ThreeColumnVerticalSection = 8,
+        /// <summary>
+        /// Two columns, left 2/3, right 1/3 + one vertical section column
+        /// </summary>
+        TwoColumnLeftVerticalSection = 9,
+        /// <summary>
+        /// Two columns, left 1/3, right 2/3 + one vertical section column
+        /// </summary>
+        TwoColumnRightVerticalSection = 10
+#endif
     }
 #endif
 }