+++ Core/OfficeDevPnP.Core/Pages/ClientSideText.cs
@@ -112,11 +112,14 @@ namespace OfficeDevPnP.Core.Pages
                     ZoneIndex = this.Section.Order,
                     SectionIndex = this.Column.Order,
                     SectionFactor = this.Column.ColumnFactor,
+#if !SP2019
+                    LayoutIndex = this.Column.LayoutIndex,
+#endif
                     ControlIndex = controlIndex,
                 },
                 Emphasis = new ClientSideSectionEmphasis()
                 {
-                    ZoneEmphasis = this.Section.ZoneEmphasis,
+                    ZoneEmphasis = this.Column.VerticalSectionEmphasis.HasValue ? this.Column.VerticalSectionEmphasis.Value : this.Section.ZoneEmphasis,
                 },
                 EditorType = "CKEditor"
             };
@@ -167,9 +170,9 @@ namespace OfficeDevPnP.Core.Pages
 #endif
             return html.ToString();
         }
-        #endregion
+#endregion
 
-        #region Internal and private methods
+                    #region Internal and private methods
         internal override void FromHtml(IElement element)
         {
             base.FromHtml(element);
@@ -207,7 +210,7 @@ namespace OfficeDevPnP.Core.Pages
             this.spControlData = JsonConvert.DeserializeObject<ClientSideTextControlData>(element.GetAttribute(CanvasControl.ControlDataAttribute), jsonSerializerSettings);
             this.controlType = this.spControlData.ControlType;
         }
-        #endregion
+                    #endregion
     }
 #endif
-}
+                }