+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 ### Added
 
+- Vertical section support (read, update and create) for the modern page API
 - Add extension to create team site with no group using SPSiteManager #2364 [NicolajHedeager]
 - Improved image extraction when extracting modern pages: header image and images linked by quick links webparts are now included #2336 [heinrich-ulbricht]
 - Added support for page header images which reside outside current site collection #2360 [gautamdsheth]